add revamped indexing type, type partial eqs, printd, LLVM string indexing

+++ src/compiler/value/internals.rs
@@ -8,6 +8,20 @@ use crate::compiler::{Compiler, CompileResult};
 
 use super::TypeLayout;
 
+macro_rules! std_map {
+    ($($l:literal: $i:ident),+) => {
+        fn lookup(&self, name: &str) -> bool {
+            matches!(name, $($l)|+)
+        }
+
+        fn register(&self, name: &str, builder: Builder<'ctx>, fun: FunctionValue<'ctx>) -> CompileResult<'ctx, ()> {
+            match name {
+                $($l => self.$i(builder, fun)),+,
+                _ => unimplemented!()
+            }
+        }
+    }
+}
 impl<'ctx> Compiler<'ctx> {
     /// [`super::TypeLayout::Str`]
     pub(super) fn string_type(&self) -> StructType<'ctx> {
@@ -31,29 +45,51 @@ impl<'ctx> Compiler<'ctx> {
             TypeLayout::Int.fn_type(self, &[self.ctx.i8_type().ptr_type(Default::default()).into()], false),
         )?;
     
-        builder.build_call(puts, &[buf.into()], "puts_call");
+        builder.build_call(puts, &[buf.into()], "");
         builder.build_return(None);
     
         Ok(())
     }
 
+    fn std_printd(&self, builder: Builder<'ctx>, fun: FunctionValue<'ctx>) -> CompileResult<'ctx, ()> {
+        let p0 = fun.get_first_param().unwrap();
+    
+        let printf = self.import_fun("printf",
+            TypeLayout::Int.fn_type(self, &[
+                self.ctx.i8_type().ptr_type(Default::default()).into()
+            ], true),
+        )?;
+    
+        let string = self.ctx.const_string(b"%d", true);
+        let alloca = builder.build_alloca(string.get_type(), "alloca");
+        builder.build_store(alloca, string);
+
+        builder.build_call(printf, &[alloca.into(), p0.into()], "");
+        builder.build_return(None);
+    
+        Ok(())
+    }
+
+    // HACK
+    std_map! {
+        "print": std_print,
+        "printd": std_printd
+    }
+
     pub fn import_fun(&self, s: &str, ty: FunctionType<'ctx>) -> CompileResult<'ctx, FunctionValue<'ctx>> {
         let fun = self.module.get_function(s).unwrap_or_else(|| self.module.add_function(s, ty, None));
 
-        if fun.count_basic_blocks() < 1 {
-            // HACK
-            if s == "print" {
-                let builder = self.ctx.create_builder();
-                let bb = self.ctx.append_basic_block(fun, "body");
-                builder.position_at_end(bb);
-    
-                self.std_print(builder, fun)?;
-    
-                if !fun.verify(true) {
-                    // SAFETY: Not used after.
-                    unsafe { fun.delete() }
-                    panic!("could not import fun '{s}'");
-                }
+        if self.lookup(s) && fun.count_basic_blocks() < 1 {
+            let builder = self.ctx.create_builder();
+            let bb = self.ctx.append_basic_block(fun, "body");
+            builder.position_at_end(bb);
+
+            self.register(s, builder, fun)?;
+
+            if !fun.verify(true) {
+                // SAFETY: Not used after.
+                unsafe { fun.delete() }
+                panic!("could not import fun '{s}'");
             }
         }
